Rename `addCustomerSaga` file to `saveCustomerSaga`
