+++ src/store/saga.tsx
@@ -1,5 +1,5 @@
 import { all } from "redux-saga/effects";
-import { watchSaveCustomer } from "./sagas/customers/addCustomerSaga";
+import { watchSaveCustomer } from "./sagas/customers/saveCustomerSaga";
 import { watchSyncCustomers } from "./sagas/customers/syncCustomersSaga";
 
 export default function* rootSaga() {
